Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 
 int main() {
-    int x;
-    scanf("%d", &x);
+    float x;
+    scanf("%f", &x);
     x = 3.14 * x * x;
-    printf("Area: %d", x);
+    printf("Area: %f", x);
     return 0;
 }
